Refactoring: split clients in separate files
Squashed commit of the following:

commit ae5e05c83133f4bb28c753b7ea0d9ce16dfe642f
Author: Aurélien Grosdidier <aurelien.grosdidier@gmail.com>
Date:   Fri Dec 7 07:42:59 2018 +0100

    Improved pylint score

commit fef72410ea81616a83fd7ed74699d174821c73cd
Author: Aurélien Grosdidier <aurelien.grosdidier@gmail.com>
Date:   Fri Dec 7 07:12:38 2018 +0100

    Reorganized clients in separate files

commit 58d5f607acef989b7ab4b3dea2ba87aff268f4fd
Author: Aurélien Grosdidier <aurelien.grosdidier@gmail.com>
Date:   Fri Dec 7 07:12:19 2018 +0100

    Ignores .mypy directory
