+++ .gitignore
@@ -8,3 +8,4 @@ feedspora.yml
 .*.swp
 .cache
 MANIFEST
+.mypy_cache
