+++ tests/mkrichtext_test.py
@@ -5,7 +5,7 @@ import string
 
 import pytest
 
-from feedspora.feedspora_runner import mkrichtext
+from feedspora.generic_client import GenericClient
 
 TBD = 'tests/'
 
@@ -14,17 +14,20 @@ TBD = 'tests/'
 def testcases():
     def make_fake_keyword():
         length = random.randint(3, 10)
+
         return ''.join(
             [random.choice(string.ascii_lowercase) for i in range(length)])
 
     def make_fake_keywords():
         number = random.randint(2, 5)
+
         return [make_fake_keyword() for i in range(number)]
 
     to_return = {}
     with open(TBD + 'phrases') as f:
         for l in [l.strip() for l in f]:
             words = [x for x in l.split(' ') if len(x) > 3]
+
             if len(words) < 3:
                 continue
             keywords = random.sample(
@@ -37,13 +40,14 @@ def testcases():
         to_return[statement] = []
         # No keywords and long
         to_return[statement * 30] = []
+
     return to_return
 
 
 def test_mkrichtext_length(testcases):
     for (text, keywords) in testcases.items():
         for maxlen in range(30, 260, 10):
-            output = mkrichtext(text, keywords, maxlen)
+            output = GenericClient()._mkrichtext(text, keywords, maxlen)
             assert not len(output) > maxlen, "{} > {}" \
                 .format(len(output))
             assert not output.endswith('|'), "{} ends with sep".format(output)
@@ -51,62 +55,64 @@ def test_mkrichtext_length(testcases):
 
 def test_mkrichtext_case(testcases):
     for (text, keywords) in testcases.items():
-        output1 = mkrichtext(text, keywords, 500)
-        output2 = mkrichtext(text.upper(), keywords, 500)
+        output1 = GenericClient()._mkrichtext(text, keywords, 500)
+        output2 = GenericClient()._mkrichtext(text.upper(), keywords, 500)
         assert output1.upper() == output2.upper()
+
 
 def test_mkrichtext_tags():
     '''
     Test the proper handling and application of hashtags
     '''
     testcases = {
-                     'No Tags, No Keywords': {
-                         'keywords': [],
-                         'expected': 'No Tags, No Keywords'
-                     },
-                     'One #Tag, No Keywords': {
-                         'keywords': [],
-                         'expected': 'One #Tag, No Keywords'
-                     },
-                     'One #Tag, Unrelated Keyword': {
-                         'keywords': ['appended'],
-                         'expected': 'One #Tag, Unrelated Keyword | #appended'
-                     },
-                     'One #Tag, Duplicate Keyword': {
-                         'keywords': ['tag'],
-                         'expected': 'One #Tag, Duplicate Keyword'
-                     },
-                     'One #Tag, Keyword In Title': {
-                         'keywords': ['keyword'],
-                         'expected': 'One #Tag, #Keyword In Title'
-                     },
-                     'tag-text-embedded In Title': {
-                         'keywords': ['tag', 'embedded'],
-                         'expected': 'tag-text-embedded In Title | #tag #embedded'
-                     },
-                     'Tags at beginning and end': {
-                         'keywords': ['tags', 'end'],
-                         'expected': '#Tags at beginning and #end'
-                     },
-                     'Tags within "quotes" and parens (tricky)': {
-                         'keywords': ['quotes', 'tricky'],
-                         'expected': 'Tags within "#quotes" and parens (#tricky)'
-                     },
-                     'Tags in red/white/blue': {
-                         'keywords': ['red', 'white', 'blue'],
-                         'expected': 'Tags in #red/#white/#blue'
-                     },
-                     'Tag before comma, before period.': {
-                         'keywords': ['comma', 'period'],
-                         'expected': 'Tag before #comma, before #period.'
-                     },
-                     '#Bad-tag retained (not bad, eh?)': {
-                         'keywords': ['bad', 'eh'],
-                         'expected': '#Bad-tag retained (not #bad, #eh?)'
-                     },
-                }
+        'No Tags, No Keywords': {
+            'keywords': [],
+            'expected': 'No Tags, No Keywords'
+        },
+        'One #Tag, No Keywords': {
+            'keywords': [],
+            'expected': 'One #Tag, No Keywords'
+        },
+        'One #Tag, Unrelated Keyword': {
+            'keywords': ['appended'],
+            'expected': 'One #Tag, Unrelated Keyword | #appended'
+        },
+        'One #Tag, Duplicate Keyword': {
+            'keywords': ['tag'],
+            'expected': 'One #Tag, Duplicate Keyword'
+        },
+        'One #Tag, Keyword In Title': {
+            'keywords': ['keyword'],
+            'expected': 'One #Tag, #Keyword In Title'
+        },
+        'tag-text-embedded In Title': {
+            'keywords': ['tag', 'embedded'],
+            'expected': 'tag-text-embedded In Title | #tag #embedded'
+        },
+        'Tags at beginning and end': {
+            'keywords': ['tags', 'end'],
+            'expected': '#Tags at beginning and #end'
+        },
+        'Tags within "quotes" and parens (tricky)': {
+            'keywords': ['quotes', 'tricky'],
+            'expected': 'Tags within "#quotes" and parens (#tricky)'
+        },
+        'Tags in red/white/blue': {
+            'keywords': ['red', 'white', 'blue'],
+            'expected': 'Tags in #red/#white/#blue'
+        },
+        'Tag before comma, before period.': {
+            'keywords': ['comma', 'period'],
+            'expected': 'Tag before #comma, before #period.'
+        },
+        '#Bad-tag retained (not bad, eh?)': {
+            'keywords': ['bad', 'eh'],
+            'expected': '#Bad-tag retained (not #bad, #eh?)'
+        },
+    }
+
     for input in testcases:
         keywords = testcases[input]['keywords']
         expected = testcases[input]['expected']
-        output = mkrichtext(input, keywords, 500)
+        output = GenericClient()._mkrichtext(input, keywords, 500)
         assert output == expected
